GEODE-5135: Refactor AbstractRegionMap dependencies
`BaseRegionMap` will be an abstract base class for `AbstractRegionMap`
   and `ProxyRegionMap`. Will move common functionality into this class.
   Aims to remove dependency between `AbstractRegionMap` and `ProxyRegionMap`
Move ARMLockTestHook to RegionMap to remove circular dependency
   while keeping the impact as small as possible.
Move `forceInvalidateEvent`, `shouldInvokeCallbacks` and
   `switchEventOwnerAndOriginRemote` to `BaseRegionMap` to reduce
   dependency between `ProxyRegionMap` and `AbstractRegionMap`
Move `createCallbackEvent` from `AbstractRegionMap` to `EntryEventImpl`
   This reduces dependecy between `AbstractRegionMap` and `ProxyRegionMap`
